scripts/requirements.txt
# TODO: Clean up requirements.txt
numpy==2.3.4
opencv_python>=4.11
Pillow==12.0.0
python-telegram-bot==22.5
